requirements.txt
alembic==1.11.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
clickclick==20.10.2
comm==0.1.4
connexion==2.14.2
decorator==5.1.1
dnspython==2.3.0
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema>=4.0.0,<5.0.0
jsonschema-specifications
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow
marshmallow-sqlalchemy==0.29.0
packaging==23.1
pathlib==1.0.1
platformdirs==3.10.0
psutil==5.9.5
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg-pool==3.1.7
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv
pytz==2023.3
PyYAML==6.0.1
referencing
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
swagger-ui-bundle==0.0.9
tornado
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1
wtforms-validators==1.0.0

